put the open and closed scanner in different files
